upgrade a few more modules, considering more macroness
The proposed macros is currently on api::test. I'm unsure if this will improve or harm by-hand maintenance. It removes nearly all the boilerplate, but doesn't quite look like the struct definitions that it expands to. Should it be passed whole struct definitions? It's also unclear how this will interact with more complicated cases, like requests that have no request type or require other type definitions to function.